[clean,add]
ファイル整理
プレイヤーリストフォーム追加
config フォームほぼ完成
