+++ scripts/util/Teleport.ts
@@ -1,5 +1,5 @@
 import { Player } from "@minecraft/server";
-import { ShuffleArray } from "./util";
+import { Execute, ShuffleArray } from "./util";
 
 const EachTeleport = (playerA: Player, playerB: Player, playerC?: Player): void => {
     const {
@@ -29,8 +29,8 @@ const EachTeleport = (playerA: Player, playerB: Player, playerC?: Player): void 
 
 export const ShuffleTeleport = (players: Array<Player>): void => {
     const shuffled = ShuffleArray(players, 20);
+    Execute((player) => player.onScreenDisplay.setActionBar("§cswapping..."));
     shuffled.forEach((player, index) => {
-        player.onScreenDisplay.setActionBar("§cswapping...");
         if (index % 2 != 0 || shuffled.length == index + 1) return;
         if (shuffled.length == index + 3) {
             EachTeleport(shuffled[index], shuffled[index + 1], shuffled[index + 2]);
